// packages/client/src/api/api.ts
 export interface SafeTransaction {
     /**
      * 
      * @type {string}
      * @memberof SafeTransaction
      */
     'to': string;
     /**
      * 
      * @type {string}
      * @memberof SafeTransaction
      */
     'value': string;
     /**
      * 
      * @type {string}
      * @memberof SafeTransaction
      */
     'data': string;
     /**
      * 
      * @type {string}
      * @memberof SafeTransaction
      */
     'operation': string;
     /**
      * 
      * @type {string}
      * @memberof SafeTransaction
      */
     'safeTxGas': string;
     /**
      * 
      * @type {string}
      * @memberof SafeTransaction
      */
     'baseGas': string;
     /**
      * 
      * @type {string}
      * @memberof SafeTransaction
      */
     'gasPrice': string;
     /**
      * 
      * @type {string}
      * @memberof SafeTransaction
      */
     'gasToken': string;
     /**
      * 
      * @type {string}
      * @memberof SafeTransaction
      */
     'refundReceiver': string;
     /**
      * 
      * @type {string}
      * @memberof SafeTransaction
      */
     'nonce': string;
 }